move common things to `generic::OverWindow`
