+++ src/frontend/src/optimizer/plan_node/generic/mod.rs
@@ -46,6 +46,8 @@ mod share;
 pub use share::*;
 mod dedup;
 pub use dedup::*;
+mod over_window;
+pub use over_window::*;
 
 pub trait GenericPlanRef {
     fn schema(&self) -> &Schema;
