+++ src/frontend/src/optimizer/plan_node/logical_over_agg.rs
@@ -16,90 +16,21 @@ use std::fmt;
 
 use fixedbitset::FixedBitSet;
 use itertools::Itertools;
-use risingwave_common::catalog::{Field, Schema};
 use risingwave_common::error::{ErrorCode, Result};
 use risingwave_common::types::DataType;
-use risingwave_common::util::sort_util::{ColumnOrder, ColumnOrderDisplay};
+use risingwave_common::util::sort_util::ColumnOrder;
 use risingwave_expr::function::window::{Frame, FrameBound, WindowFuncKind};
 
+use super::generic::{OverWindow, PlanWindowFunction};
 use super::{
     gen_filter_and_pushdown, ColPrunable, ExprRewritable, LogicalProject, PlanBase, PlanRef,
     PlanTreeNodeUnary, PredicatePushdown, ToBatch, ToStream,
 };
-use crate::expr::{Expr, ExprImpl, InputRef, InputRefDisplay, WindowFunction};
+use crate::expr::{Expr, ExprImpl, InputRef, WindowFunction};
 use crate::optimizer::plan_node::{
     ColumnPruningContext, PredicatePushdownContext, RewriteStreamContext, ToStreamContext,
 };
-use crate::utils::{ColIndexMapping, ColIndexMappingRewriteExt, Condition};
-
-/// Rewritten version of [`WindowFunction`] which uses `InputRef` instead of `ExprImpl`.
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct PlanWindowFunction {
-    pub kind: WindowFuncKind,
-    pub return_type: DataType,
-    pub args: Vec<InputRef>,
-    pub partition_by: Vec<InputRef>,
-    pub order_by: Vec<ColumnOrder>,
-    pub frame: Option<Frame>,
-}
-
-struct PlanWindowFunctionDisplay<'a> {
-    pub window_function: &'a PlanWindowFunction,
-    pub input_schema: &'a Schema,
-}
-
-impl<'a> std::fmt::Debug for PlanWindowFunctionDisplay<'a> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let window_function = self.window_function;
-        if f.alternate() {
-            f.debug_struct("WindowFunction")
-                .field("kind", &window_function.kind)
-                .field("return_type", &window_function.return_type)
-                .field("partition_by", &window_function.partition_by)
-                .field("order_by", &window_function.order_by)
-                .field("frame", &window_function.frame)
-                .finish()
-        } else {
-            write!(f, "{}() OVER(", window_function.kind)?;
-
-            let mut delim = "";
-            if !window_function.partition_by.is_empty() {
-                delim = " ";
-                write!(
-                    f,
-                    "PARTITION BY {}",
-                    window_function
-                        .partition_by
-                        .iter()
-                        .format_with(", ", |input_ref, f| {
-                            f(&InputRefDisplay {
-                                input_ref,
-                                input_schema: self.input_schema,
-                            })
-                        })
-                )?;
-            }
-            if !window_function.order_by.is_empty() {
-                write!(
-                    f,
-                    "{delim}ORDER BY {}",
-                    window_function.order_by.iter().format_with(", ", |o, f| {
-                        f(&ColumnOrderDisplay {
-                            column_order: o,
-                            input_schema: self.input_schema,
-                        })
-                    })
-                )?;
-            }
-            if let Some(frame) = &window_function.frame {
-                write!(f, "{delim}{}", frame)?;
-            }
-            f.write_str(")")?;
-
-            Ok(())
-        }
-    }
-}
+use crate::utils::{ColIndexMapping, Condition};
 
 /// `LogicalOverAgg` performs `OVER` window aggregates ([`WindowFunction`]) to its input.
 ///
@@ -107,34 +38,14 @@ impl<'a> std::fmt::Debug for PlanWindowFunctionDisplay<'a> {
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub struct LogicalOverAgg {
     pub base: PlanBase,
-    pub window_functions: Vec<PlanWindowFunction>,
-    input: PlanRef,
+    core: OverWindow<PlanRef>,
 }
 
 impl LogicalOverAgg {
-    fn new(window_functions: Vec<PlanWindowFunction>, input: PlanRef) -> Self {
-        let ctx = input.ctx();
-        let mut schema = input.schema().clone();
-        window_functions.iter().for_each(|func| {
-            schema.fields.push(Field::with_name(
-                func.return_type.clone(),
-                func.kind.to_string(),
-            ));
-        });
-
-        let logical_pk = input.logical_pk().to_vec();
-
-        let mapping = ColIndexMapping::identity_or_none(input.schema().len(), schema.len());
-        let fd_set = input.functional_dependency().clone();
-        let fd_set = mapping.rewrite_functional_dependency_set(fd_set);
-
-        let base = PlanBase::new_logical(ctx, schema, logical_pk, fd_set);
-
-        Self {
-            base,
-            window_functions,
-            input,
-        }
+    fn new(calls: Vec<PlanWindowFunction>, input: PlanRef) -> Self {
+        let core = OverWindow::new(calls, input);
+        let base = PlanBase::new_logical_with_core(&core);
+        Self { base, core }
     }
 
     pub fn create(
@@ -273,6 +184,10 @@ impl LogicalOverAgg {
         })
     }
 
+    pub fn window_functions(&self) -> &[PlanWindowFunction] {
+        &self.core.window_functions
+    }
+
     #[must_use]
     fn rewrite_with_input_and_window(
         &self,
@@ -302,11 +217,11 @@ impl LogicalOverAgg {
 
 impl PlanTreeNodeUnary for LogicalOverAgg {
     fn input(&self) -> PlanRef {
-        self.input.clone()
+        self.core.input.clone()
     }
 
     fn clone_with_input(&self, input: PlanRef) -> Self {
-        Self::new(self.window_functions.clone(), input)
+        Self::new(self.core.window_functions.clone(), input)
     }
 
     #[must_use]
@@ -316,7 +231,7 @@ impl PlanTreeNodeUnary for LogicalOverAgg {
         input_col_change: ColIndexMapping,
     ) -> (Self, ColIndexMapping) {
         let over_agg =
-            self.rewrite_with_input_and_window(input, &self.window_functions, input_col_change);
+            self.rewrite_with_input_and_window(input, self.window_functions(), input_col_change);
         // change the input columns index will not change the output column index
         let out_col_change = ColIndexMapping::identity(over_agg.schema().len());
         (over_agg, out_col_change)
@@ -327,17 +242,7 @@ impl_plan_tree_node_for_unary! { LogicalOverAgg }
 
 impl fmt::Display for LogicalOverAgg {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let mut builder = f.debug_struct("LogicalOverAgg");
-        let window_funcs_display = self
-            .window_functions
-            .iter()
-            .map(|func| PlanWindowFunctionDisplay {
-                window_function: func,
-                input_schema: self.input.schema(),
-            })
-            .collect::<Vec<_>>();
-        builder.field("window_functions", &window_funcs_display);
-        builder.finish()
+        self.core.fmt_with_name(f, "LogicalOverAgg")
     }
 }
 
@@ -360,7 +265,7 @@ impl ColPrunable for LogicalOverAgg {
                 .filter(|&&index| index >= input_cnt)
                 .map(|&index| {
                     let index = index - input_cnt;
-                    let window_function = self.window_functions[index].clone();
+                    let window_function = self.window_functions()[index].clone();
                     tmp.extend(window_function.args.iter().map(|x| x.index()));
                     tmp.extend(window_function.partition_by.iter().map(|x| x.index()));
                     tmp.extend(window_function.order_by.iter().map(|x| x.column_index));
@@ -414,7 +319,7 @@ impl PredicatePushdown for LogicalOverAgg {
         ctx: &mut PredicatePushdownContext,
     ) -> PlanRef {
         let mut window_col = FixedBitSet::with_capacity(self.schema().len());
-        window_col.insert_range(self.input.schema().len()..self.schema().len());
+        window_col.insert_range(self.core.input.schema().len()..self.schema().len());
         let (window_pred, other_pred) = predicate.split_disjoint(&window_col);
         gen_filter_and_pushdown(self, window_pred, other_pred, ctx)
     }
