+++ src/frontend/src/optimizer/plan_node/mod.rs
@@ -722,7 +722,7 @@ pub use logical_join::LogicalJoin;
 pub use logical_limit::LogicalLimit;
 pub use logical_multi_join::{LogicalMultiJoin, LogicalMultiJoinBuilder};
 pub use logical_now::LogicalNow;
-pub use logical_over_agg::{LogicalOverAgg, PlanWindowFunction};
+pub use logical_over_agg::LogicalOverAgg;
 pub use logical_project::LogicalProject;
 pub use logical_project_set::LogicalProjectSet;
 pub use logical_scan::LogicalScan;
